tests/wpt/web-platform-tests/css/css-fonts/animations/font-size-interpolation-002.html

<!DOCTYPE html>
<meta charset="UTF-8">
<title>font-size interpolation</title>
<link rel="help" href="https://drafts.csswg.org/css-fonts-3/#propdef-font-size">
<meta name="assert" content="font-size supports animation as length">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/interpolation-testcommon.js"></script>

<style>
.container {
  font-size: 10px;
}
.target {
  display: inline-block;
  font: 100px sans-serif;
  font-size: 20px;
}
.expected {
  color: green;
  margin-right: 30px;
}
</style>

<body>
<template id="target-template">
  <span class="container">
    <div class="target">TT</div>
  </span>
</template>
</body>

<script>
// Test the 'unset' case.
test_interpolation({
  property: 'font-size',
  from: 'unset',
  to: '20px'
}, [
  {at: -2, expect: '0px'},  // CSS font-size can't be negative.
  {at: -0.3, expect: '7px'},
  {at: 0, expect: '10px'},
  {at: 0.3, expect: '13px'},
  {at: 0.6, expect: '16px'},
  {at: 1, expect: '20px'},
  {at: 1.5, expect: '25px'},
]);
</script>
